requirements.txt
einops==0.6.1
hydra-core==1.3.2
ipython==8.14.0
matplotlib==3.5.1
numpy==1.22.4
omegaconf==2.0.6
pandas==1.3.4
partitura==1.2.1
pytorch_lightning==1.9.2
scikit_learn==1.3.0
scipy==1.7.3
torch==1.10.2
tqdm==4.64.0
